Introduce shared adapter
